refact: 도메인 데이터 별이 아닌 게임 흐름 별 Service 클래스 작성
