+++ gotcha-socket/src/main/java/socket_server/domain/game/service/GameBroadCaster.java
@@ -1,0 +1,58 @@
+package socket_server.domain.game.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Service;
+import socket_server.common.config.RedisMessage;
+import socket_server.common.util.JsonSerializer;
+import socket_server.domain.game.dto.AISaysRes;
+import socket_server.domain.game.dto.GameRes;
+import socket_server.domain.game.enumType.GameEventType;
+import socket_server.domain.room.dto.EventRes;
+import socket_server.domain.room.dto.EventType;
+
+import java.time.LocalDateTime;
+
+import static socket_server.common.constants.WebSocketConstants.GAME_PREFIX;
+import static socket_server.common.constants.WebSocketConstants.ROOM_EVENT;
+
+@Service
+@RequiredArgsConstructor
+public class GameBroadCaster {
+
+    private final JsonSerializer jsonSerializer;
+    private final RedisTemplate<String, Object> objectRedisTemplate;
+
+    public void broadcastStartEvent(String userUuid, String roomId, AISaysRes aiSaysRes) {
+        EventRes eventRes = new EventRes(
+                EventType.START,
+                aiSaysRes,
+                LocalDateTime.now()
+        );
+
+        RedisMessage redisMessage = new RedisMessage(
+                userUuid,
+                ROOM_EVENT + roomId,
+                jsonSerializer.serialize(eventRes)
+        );
+
+        objectRedisTemplate.convertAndSend(ROOM_EVENT + roomId, jsonSerializer.serialize(redisMessage));
+    }
+
+    public void broadcastGameEvent(String senderUuid, String roomId, GameEventType gameEventType, Object data) {
+        GameRes gameRes = new GameRes(
+                gameEventType,
+                data,
+                LocalDateTime.now());
+
+        objectRedisTemplate.convertAndSend(
+                GAME_PREFIX + roomId,
+                new RedisMessage(
+                        senderUuid,
+                        GAME_PREFIX + roomId,
+                        jsonSerializer.serialize(gameRes)
+                )
+        );
+
+    }
+}
