--- gotcha-socket/src/main/java/socket_server/domain/game/service/GamePlayerService.java
@@ -1,25 +1,0 @@
-package socket_server.domain.game.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import socket_server.domain.game.model.GamePlayer;
-import socket_server.domain.game.repository.GameRepository;
-import socket_server.domain.room.model.RoomUserInfo;
-import socket_server.domain.room.repository.RoomUserRepository;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class GamePlayerService {
-    /**
-     * 도메인 데이터 GamePlayer 처리 담당.
-     */
-    private final RoomUserRepository roomUserRepository;
-
-    public List<GamePlayer> getGamePlayersFromRoom(String roomId) {
-        return roomUserRepository.findUsersByRoomId(roomId).stream().map(RoomUserInfo::toGamePlayer).toList();
-    }
-
-
-}
