+++ gotcha-socket/src/main/java/socket_server/domain/room/handler/GameStartHandler.java
@@ -5,7 +5,7 @@ package socket_server.domain.room.handler;
 import gotcha_domain.auth.SecurityUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import socket_server.domain.game.service.GameService;
+import socket_server.domain.game.service.GameStartService;
 import socket_server.domain.room.dto.EventType;
 import socket_server.domain.room.dto.RoomReq;
 
@@ -13,7 +13,7 @@ import socket_server.domain.room.dto.RoomReq;
 @RequiredArgsConstructor
 public class GameStartHandler implements RoomEventHandler {
 
-    private final GameService gameService;
+    private final GameStartService gameStartService;
     @Override
     public EventType getEventType() {
         return EventType.START;
@@ -21,7 +21,7 @@ public class GameStartHandler implements RoomEventHandler {
 
     @Override
     public void handle(String roomId, SecurityUserDetails userDetails, RoomReq request) {
-        gameService.startGame(roomId, userDetails.getUuid());
+        gameStartService.startGame(roomId, userDetails.getUuid());
     }
 
 }
